requirements/local.txt
-r base.txt

# Debug toolbar
django-debug-toolbar==1.9.1